clion rocks

+++ ft_printf/libft/ft_string_of_spaces.c
@@ -8,7 +8,7 @@ char 	*ft_string_of_spaces(unsigned int size)
 	i = 0;
 	if (!(str = (char *)malloc(sizeof(char) * (size + 1))))
 		return (NULL);
-	while (i < size)
+	while ((unsigned int)i < size)
 		((char *)str)[i++] = (char)' ';
 	str[i] = '\0';
 	return (str);
